requirements.txt
absl-py==0.10.0
backcall==0.2.0
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
fsspec==0.8.4
future==0.18.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
hyperopt=0.2.5
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.46
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==0.10.0
PyYAML==5.3.1
ray==1.3.0
ray[tune]
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
scikit-learn==0.24.1
statsmodels==0.12.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0